backend/requirements.txt
convex>=0.4.0
python-dotenv>=1.0.0
ai-sdk-python==0.1.0
kagglehub>=0.2.0
pandas>=2.0.0
openai>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
faiss-cpu>=1.7.4
numpy>=1.24.0
email-validator>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
exa-py>=1.0.0
anthropic>=0.34.0
PyPDF2>=3.0.0
scikit-learn>=1.3.0
python-multipart>=0.0.6
psutil>=5.9.0
pydantic>=2.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
playwright>=1.40.0
